requirements.txt
praw==7.7.1
playwright==1.44.0
beautifulsoup4==4.12.3
httpx==0.27.0
transformers==4.41.2
torch==2.3.1
python-dotenv==1.0.1
apscheduler==3.10.4
fastapi==0.111.0
uvicorn==0.30.1
jinja2==3.1.4
rapidfuzz==3.9.3
aiosqlite==0.20.0
